+++ src/app/app.component.ts
@@ -6,6 +6,7 @@ import { NotificationService } from './components/notification/notification.serv
 import { MARKUP_TYPES } from 'src/rxcore/constants';
 import { AnnotationToolsService } from './components/annotation-tools/annotation-tools.service';
 import { RecentFilesService } from './components/recent-files/recent-files.service';
+import { UserService } from './components/user/user.service';
 import { Title } from '@angular/platform-browser';
 import { IGuiConfig } from 'src/rxcore/models/IGuiConfig';
 
@@ -21,7 +22,7 @@ export class AppComponent implements AfterViewInit {
   guiConfig$ = this.rxCoreService.guiConfig$;
   guiConfig: IGuiConfig | undefined;
   title: string = 'rasterex-viewer';
-  uiversion : string = '12.0.0.8'
+  uiversion : string = '12.1.0.1'
   numOpenFiles$ = this.rxCoreService.numOpenedFiles$;
   annotation: any;
   rectangle: any;
@@ -44,6 +45,7 @@ export class AppComponent implements AfterViewInit {
     private readonly rxCoreService: RxCoreService,
     private readonly fileGaleryService: FileGaleryService,
     private readonly notificationService: NotificationService,
+    private readonly userService: UserService,
     private titleService:Title) { }
 
   ngOnInit() {
@@ -205,7 +207,17 @@ export class AppComponent implements AfterViewInit {
             
       this.rxCoreService.guiFileLoadComplete.next();
 
+      // TODO: The settings are effective after the file is loaded completely.
+      this.userService.canUpdateAnnotation$.subscribe((canUpdate) => {
+        // By setting the markup lock, operations such as dragging the markup with the mouse are prohibited.
+        RXCore.lockMarkup(!canUpdate);
+      });
+
+      this.userService.canViewAnnotation$.subscribe((canView) => {
+        //RXCore.hideMarkUp();
+      });
       
+     
       
 
     });